move cfx_filter trait to its own file
