+++ crates/client/src/rpc/traits/mod.rs
@@ -2,7 +2,7 @@
 // Conflux is free software and distributed under GNU General Public License.
 // See http://www.gnu.org/licenses/
 
-pub use cfx_space::{cfx, debug, pool, pos, pubsub, test, trace};
+pub use cfx_space::{cfx, cfx_filter, debug, pool, pos, pubsub, test, trace};
 
 pub mod cfx_space;
 pub mod eth_space;
